+++ Moth/LLVM/Metadata/ParamType.cs
@@ -2,6 +2,6 @@ namespace Moth.LLVM.Metadata;
 
 public struct ParamType
 {
-    public ulong typeref_table_index;
-    public ulong typeref_table_length;
+    public uint typeref_table_index;
+    public uint typeref_table_length;
 }
